A step, a fall, a thrown blade, a fistful of dirt
Four more of Bushido Blade's controls. Double-tapping a move control
dashes - a burst step on the dodge's spring, detected at the input
source, stance kept. C lies down flat on purpose and presses again to
rise: lyingDown folds into downed(), so the crawl, the side roll, the
prone capsule and prone swings all come free, and the topple integrator
grows the one new piece - a rise branch that unwinds the tilt. T throws
what the hand holds: armed, the blade itself flies live for 0.7 s,
blockable, jab-grade on hit, and gone from the hand either way;
empty-handed and crouched (or on the floor), a fistful of dirt instead -
unblockable, and two seconds of browned-over screen for a human or
refused reads for the bot.

The dash/lay-down/hurl bits overflowed the two-byte input packet, so the
wire format widens to u32 (19 bits used, 13 spare): kReplayVersion 2,
kProtocol 6. Dirt in flight is hashed sim state. The bot dashes out of
defensive reads, gambles the occasional mid-range throw, and goes
helpless when blinded, which is the point of blinding it.

+++ src/netplay/replay.cpp
@@ -82,7 +82,7 @@ bool Recorder::open(const char* path, const MatchSetup& setup) {
     return true;
 }
 
-void Recorder::step(std::uint16_t inputA, std::uint16_t inputB,
+void Recorder::step(std::uint32_t inputA, std::uint32_t inputB,
                     std::uint32_t checksum) {
     if (!m_out.is_open()) {
         return;
@@ -164,7 +164,7 @@ bool Replayer::next(PlayerInput inputs[2], std::uint32_t& checksum) {
     if (!m_in.is_open()) {
         return false;
     }
-    std::uint16_t bits[2] = {};
+    std::uint32_t bits[2] = {};
     m_in.read(reinterpret_cast<char*>(bits), sizeof bits);
     readU32(m_in, checksum);
     if (!m_in) {
